waitlist_backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
gspread==6.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
pydantic==2.5.3
